refactor: メソッドの移動
