+++ dev/20190130/k2works/src/index.js
@@ -2,10 +2,15 @@ import "./contents.css";
 import "./fizzbuzz_controller";
 import "./fizzbuzz_view";
 import $ from "jquery";
-import {execute, fizzBuzz, iterate} from "./fizzbuzz_model";
+import {execute} from "./fizzbuzz_model";
 import {FIZZ_BUZZ_MAX_COUNT as MAX_COUNT} from "./constants";
-
-let fizzBuzzUpDownCount = 0;
+import {
+  fizzBuzzCall, fizzBuzzDown, fizzBuzzIterate, fizzBuzzPrint, fizzBuzzUp,
+  showFizzBuzzCall,
+  showFizzBuzzIterate,
+  showFizzBuzzPrint,
+  showFizzBuzzUpDown
+} from "./fizzbuzz_controller";
 
 execute(MAX_COUNT);
 bindActions();
@@ -45,136 +50,4 @@ function bindActions() {
       menu.on("click", hash[key]);
     });
   });
-}
-
-function showFizzBuzzCall() {
-  showFizzBuzz({
-    "call": "block",
-    "print": "none",
-    "updown": "none",
-    "iterate": "none"
-  });
-}
-
-function showFizzBuzzPrint() {
-  showFizzBuzz({
-    "call": "none",
-    "print": "block",
-    "updown": "none",
-    "iterate": "none"
-  });
-}
-
-function showFizzBuzzUpDown() {
-  showFizzBuzz({
-    "call": "none",
-    "print": "none",
-    "updown": "block",
-    "iterate": "none"
-  });
-}
-
-function showFizzBuzzIterate() {
-  showFizzBuzz({
-    "call": "none",
-    "print": "none",
-    "updown": "none",
-    "iterate": "block"
-  });
-}
-
-function showFizzBuzz(
-  state = {
-    "call": "none",
-    "print": "none",
-    "updown": "none",
-    "iterate": "none"
-  }
-) {
-  $("#fizz-buzz-component__article--call").css("display", state["call"]);
-  $("#fizz-buzz-component__article--print").css("display", state["print"]);
-  $("#fizz-buzz-component__article--up-down").css("display", state["updown"]);
-  $("#fizz-buzz-component__article--iterate").css("display", state["iterate"]);
-}
-
-function fizzBuzzCall() {
-  const value = $("#fizz-buzz-component__input--call").val();
-  if (isNaN(value)) {
-    alert("数字を入力してください。");
-  } else {
-    alert(fizzBuzz(value));
-  }
-}
-
-function fizzBuzzPrint() {
-  const value = $("#fizz-buzz-component__input--print").val();
-  const target = $("#fizz-buzz-component__print--message");
-  if (isNaN(value)) {
-    target.html("数字を入力してください。");
-  } else {
-    target.html(fizzBuzz(value));
-  }
-}
-
-function fizzBuzzDown() {
-  if (fizzBuzzUpDownCount === 0) {
-    fizzBuzzUpDownCount = 0;
-  } else {
-    fizzBuzzUpDownCount -= 1;
-  }
-  const target = $("#fizz-buzz-component__up-down--message");
-  const count = fizzBuzzUpDownCount;
-  target.html(fizzBuzz(count));
-}
-
-function fizzBuzzUp() {
-  fizzBuzzUpDownCount += 1;
-  const target = $("#fizz-buzz-component__up-down--message");
-  const count = fizzBuzzUpDownCount;
-  const value = fizzBuzz(count);
-  target.html(value);
-}
-
-function fizzBuzzIterate() {
-  const target = $("#fizz-buzz-component__iterate--result");
-  let count = $("#fizz-buzz-component__input--iterate").val();
-  if (count > MAX_COUNT) {
-    return target.html(`<strong>件数は${MAX_COUNT}までです。</strong>`);
-  }
-  let array = iterate(parseInt(count));
-  target.html(renderHtmlTable(array));
-}
-
-function renderHtmlTable(array) {
-  let rowCount = 1;
-  let html = "<table>";
-
-  html = html + "<thead>";
-  [...Array(10).keys()].forEach((v) => {
-    html = html + `<th>${v + 1}</th>`;
-  });
-  html = html + "</thead>";
-
-  html = html + "<tbody>";
-
-  html = html + "<tr>";
-  array.forEach((value) => {
-    if (rowCount > 10) {
-      html = html + "<tr>";
-    }
-
-    value = `<td>${value}</td>`;
-    html = html + value;
-
-    if (rowCount === 10) {
-      html = html + "</tr>";
-      rowCount = 1;
-    } else {
-      rowCount += 1;
-    }
-  });
-
-  html = html + "</tbody>";
-  html = html + "</table>";
-  return html;
 }
